% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{assemble_dashboard}
\alias{assemble_dashboard}
\title{Combine shinydashboard elements}
\usage{
assemble_dashboard(
  SD_header_in = get_dashboard_header(),
  SD_sidebar_in = get_dashboard_sidebar(),
  SD_body_in = get_dashboard_body(),
  SD_footer_in = get_dashboard_footer()
)
}
\arguments{
\item{SD_header_in}{output from shinydashboard::dashboardHeader}

\item{SD_sidebar_in}{output from shinydashboard::dashboardSidebar()}

\item{SD_body_in}{output from shinydashboard::dashboardBody()}

\item{SD_footer_in}{output from shinydashboard::dashboardFooter()}
}
\value{
shinydashboard object to include in fluid_row
}
\description{
convenience function for combining subcomponents of the shinydashboard
}
