% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_CBR_footer_ui.R
\name{draw_CBR_footer_ui}
\alias{draw_CBR_footer_ui}
\title{Title}
\usage{
draw_CBR_footer_ui()
}
\value{
\code{shiny.tag} that is a \link[shinydashboardPlus:dashboardFooter]{shinydashboardPlus::dashboardFooter}
}
\description{
Title
}
\examples{
draw_CBR_footer_ui()

}
