% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_dashboard_header_ui.R
\name{draw_dashboard_header_ui}
\alias{draw_dashboard_header_ui}
\title{Title}
\usage{
draw_dashboard_header_ui()
}
\value{
shinydashboard header panel shinyUI code (i.e., within \code{shiny::tagList()})
}
\description{
Title
}
\examples{
draw_dashboard_header_ui()

}
