% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_details_sel_data_ui.R
\name{draw_details_sel_data_ui}
\alias{draw_details_sel_data_ui}
\title{Title}
\usage{
draw_details_sel_data_ui()
}
\value{
a \code{shiny.tag} with a \code{tags$details} object
}
\description{
Title
}
\examples{
draw_details_sel_data_ui()

}
