% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_est_more_info_div_ui.R
\name{draw_est_more_info_div_ui}
\alias{draw_est_more_info_div_ui}
\title{create ui for estimate panels}
\usage{
draw_est_more_info_div_ui()
}
\value{
a \code{shiny.tag} object inside a \code{shinydashboardPlus::box()} footer
}
\description{
create ui for estimate panels
}
\examples{
draw_est_more_info_div_ui()

}
