% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_estimates_panel_ui.R
\name{draw_estimates_panel_ui}
\alias{draw_estimates_panel_ui}
\title{create ui for estimate panels}
\usage{
draw_estimates_panel_ui()
}
\value{
a \code{shiny.tag} object inside a `shinydashboardPlus::box()``
}
\description{
create ui for estimate panels
}
\examples{
draw_estimates_panel_ui()

}
