% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_ibox_ui.R
\name{draw_ibox_ui}
\alias{draw_ibox_ui}
\title{Title}
\usage{
draw_ibox_ui(box_content_in = "daily_values_box")
}
\arguments{
\item{box_content_in}{specific text referring to a predefined HTML or shiny-renderable objected in a named list}
}
\value{
HTML object displayed to display when information button 'ibutt' is clicked or in the help
}
\description{
Title
}
\details{
lookup content to place in box or well panel. If name is not recognized than an HTML
paragraph \verb{<p> placeholder </p>} is used. See https://shiny.posit.co/py/api/core/TagTypes.html
}
\examples{
draw_ibox_ui()


}
