% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_inputs_ann_summ_dt.R
\name{draw_inputs_ann_summ_dt}
\alias{draw_inputs_ann_summ_dt}
\title{Title}
\usage{
draw_inputs_ann_summ_dt(
  input_year = input$year_picker,
  columns_used = c("Year", "WYT", "barrier", "Model")
)
}
\arguments{
\item{input_year}{}

\item{col_used}{}
}
\description{
Title
}
