% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_inputs_map_schematic_col_ui.R
\name{draw_inputs_map_schematic_col_ui}
\alias{draw_inputs_map_schematic_col_ui}
\title{Title}
\usage{
draw_inputs_map_schematic_col_ui()
}
\description{
Title
}
\examples{
draw_inputs_map_schematic_col_ui()

}
