% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_inputs_star_loc_datsrc_col_ui.R
\name{draw_inputs_star_loc_datsrc_col_ui}
\alias{draw_inputs_star_loc_datsrc_col_ui}
\title{Title}
\usage{
draw_inputs_star_loc_datsrc_col_ui(init_data_source_in)
}
\arguments{
\item{init_data_source_in}{}
}
\description{
Title
}
\examples{
draw_inputs_star_loc_datsrc_col_ui(init_data_source_in="Previous year")

}
