% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_main_page_content_dynui.R
\name{draw_main_page_content_dynui}
\alias{draw_main_page_content_dynui}
\title{Title}
\usage{
draw_main_page_content_dynui(input_tab_in = "about", init_data_source_in)
}
\arguments{
\item{input_tab_in}{}
}
\value{
ui output based on \code{uiOutput} \link[htmltools:tagList]{htmltools::tagList}
}
\description{
Title
}
\examples{
draw_main_page_content_dynui()

}
