% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_met_ref_page_ui.R
\name{draw_met_ref_page_ui}
\alias{draw_met_ref_page_ui}
\title{Title}
\usage{
draw_met_ref_page_ui()
}
\value{
method and refernce page as a  \link[htmltools:tagList]{htmltools::tagList}
}
\description{
Title
}
