% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_title_w_paren_hover_text_html.R
\name{draw_title_w_paren_hover_text_html}
\alias{draw_title_w_paren_hover_text_html}
\title{Title}
\usage{
draw_title_w_paren_hover_text_html(
  title_in = "Head of Old River to Turner Cut Junction",
  paren_htext_in = "via SJR",
  modal_text_in = "Hello World"
)
}
\arguments{
\item{title_in}{bolded title}

\item{paren_htext_in}{parenthetical plain text to the right of the title w/ same format}

\item{modal_text_in}{text that appears in the model}
}
\value{
HTML code with title and hover text over a parenthetical
}
\description{
Title
}
\examples{
draw_title_w_paren_hover_text_html(
title_in="Head of Old River to Turner Cut Junction",
paren_htext_in="via SJR",
modal_text_in="Hello World")

}
