% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_upload_details_ui.R
\name{draw_upload_details_ui}
\alias{draw_upload_details_ui}
\title{Title}
\usage{
draw_upload_details_ui()
}
\value{
a \code{shiny.tag} with a \code{tags$details} object
}
\description{
Title
}
\examples{
draw_upload_details_ui()

}
