% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_glmmTMB_ests.R
\name{extract_glmmTMB_ests}
\alias{extract_glmmTMB_ests}
\title{Title}
\usage{
extract_glmmTMB_ests(
  glmmTMB_res_ls_in,
  mods_obj_nm = "HOR_TCJ_d2_mods",
  sel_data_in = sel_rows_tmp4,
  aic_avg_tb_wts_in = aic_avg_tb_wts
)
}
\arguments{
\item{glmmTMB_res_ls_in}{}

\item{mods_obj_nm}{}

\item{sel_data_in}{}

\item{aic_avg_tb_wts_in}{}
}
\description{
Title
}
