% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_mods_preds.R
\name{get_comp_model_preds}
\alias{get_comp_model_preds}
\title{Title}
\usage{
get_comp_model_preds(
  flength_in = 244,
  DOY_in = 50:53,
  years_in = 2012,
  predict_int = TRUE
)

get_comp_model_preds(
  flength_in = 244,
  DOY_in = 50:53,
  years_in = 2012,
  predict_int = TRUE
)
}
\arguments{
\item{flength_in}{}

\item{DOY_in}{}

\item{years_in}{}

\item{predict_int}{}

\item{TCJ_CHP_pred_comp_ls_in}{}
}
\description{
Title

Title
}
