% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_custom_css_txt.R
\name{get_custom_css_txt}
\alias{get_custom_css_txt}
\title{Title}
\usage{
get_custom_css_txt(css_txt_content = "sidebar")
}
\arguments{
\item{css_txt_content}{}
}
\value{
.css definitions for specific ui elements (i.e., sidebar)
}
\description{
Title
}
\examples{
get_custom_css_txt()

}
