% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{get_dashboard_header}
\alias{get_dashboard_header}
\title{Produce shinydashboard header}
\usage{
get_dashboard_header(app_name_in = golem::pkg_name(), vers_name_in = "0.0.9")
}
\arguments{
\item{app_name_in}{name for the app}

\item{vers_name_in}{version name for the app}
}
\value{
dashboardHeader object
}
\description{
Produce shinydashboard header
}
