% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtch_WYT_txt.R
\name{get_mtch_WYT_txt}
\alias{get_mtch_WYT_txt}
\title{Title}
\usage{
get_mtch_WYT_txt(intab = ann_HORbar_WYT_data)
}
\arguments{
\item{intab}{dataframe with WYT column for matching}
}
\value{
matching color associated with name attribut for color vector
}
\description{
Title
}
\examples{
WYT_cols <- utils_get_WYT_cols_vec()
WYT_cols[match(ann_HORbar_WYT_data$WYT,names(WYT_cols))]

}
