% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_overall_surv_preds.R
\name{get_overall_surv_preds}
\alias{get_overall_surv_preds}
\title{Title}
\usage{
get_overall_surv_preds(pred_tab_ls_in, predict_int = F, conf_level = 0.95)
}
\arguments{
\item{pred_tab_ls_in}{table of predictions from component model sets \code{XXXX_mod_wrap()} functions}

\item{predict_int}{whether to include release group level erroradjSE}

\item{conf_level}{interval confidence level, based on Z-stat. Defaults to 0.95 (i.e., 1.96)}
}
\value{
named list
}
\description{
Title
}
