% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_overall_surv.R
\name{get_overall_surv_sim}
\alias{get_overall_surv_sim}
\title{Logit-Normal Network Approximation for Matrices Via Monte Carlo Simulation}
\usage{
get_overall_surv_sim(MU_mat, SIGMA2_mat, conf_level = 0.95, n_samples = 1000)
}
\arguments{
\item{MU_mat}{A matrix where each row is a scenario and the 6 columns are logit-scale means in order:
\link{,1} Psi_SLJ, \link{,2} theta_HOR-TCJ|SLJ, \link{,3} Psi_MAC, \link{,4} theta_TCJ-CHP|MAC, \link{,5} theta_TCJ-CHP|TRN, \link{,6} theta_HOR-CHP|ORE}

\item{SIGMA2_mat}{A matrix of the same dimensions containing logit-scale variances.}

\item{conf_level}{A numeric value between 0 and 1 indicating the target confidence interval width (default = 0.95).}

\item{n_samples}{Number of Monte Carlo simulation iterations per scenario row (default = 100,000).}
}
\value{
A data frame containing multi-layer simulated results for each scenario row.
}
\description{
Logit-Normal Network Approximation for Matrices Via Monte Carlo Simulation
}
