% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sidebar_txt.R
\name{get_sidebar_txt}
\alias{get_sidebar_txt}
\title{Title}
\usage{
get_sidebar_txt(tab_name_in)
}
\arguments{
\item{tab_name_in}{}
}
\value{
a \code{shiny.tag} with text
}
\description{
Title
}
\examples{
get_sidebar_txt("about")

}
