% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_center_scale.R
\name{get_var_center_scale}
\alias{get_var_center_scale}
\title{Title}
\usage{
get_var_center_scale(TMB_mod_ls)
}
\arguments{
\item{TMB_mod_ls}{named list containing pre-existing TMB models}
}
\value{
dataframe with columns correponding to the scaled variable name, its center (mean), and its scale (sd)
}
\description{
Title
}
\details{
Extract and scaled variable information
}
