% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOR_CHP_comp_alt.R
\name{HOR_CHP_comp_alt}
\alias{HOR_CHP_comp_alt}
\title{Title}
\usage{
HOR_CHP_comp_alt(HOR_CHP_data_inputs_ls_in, z_scale_vars = TRUE)
}
\arguments{
\item{HOR_CHP_data_inputs_ls_in}{}

\item{z_scale_vars}{}
}
\description{
Title
}
