% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOR_CHP_get_pred.R
\name{HOR_CHP_get_pred}
\alias{HOR_CHP_get_pred}
\title{Title}
\usage{
HOR_CHP_get_pred(
  HOR_CHP_pred_comp_ls_in = HOR_CHP_pred_comp_ls,
  xpred_tmp_in = xpred_tmp,
  avg_onlyTF = TRUE
)
}
\arguments{
\item{HOR_CHP_pred_comp_ls_in}{}

\item{xpred_tmp_in}{}

\item{avg_onlyTF}{}
}
\description{
Title
}
