% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOR_CHP_mod_wrap.R
\name{HOR_CHP_mod_wrap}
\alias{HOR_CHP_mod_wrap}
\title{Title}
\usage{
HOR_CHP_mod_wrap(
  flength_in = 240,
  DOY_in = 50:100,
  years_in = NULL,
  CVhelp_dat_w_in = CVhelp_dat_w
)
}
\arguments{
\item{flength_in}{}

\item{DOY_in}{}

\item{years_in}{}
}
\description{
Title
}
