% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOR_mod_wrap.R
\name{HOR_mod_wrap}
\alias{HOR_mod_wrap}
\title{Title}
\usage{
HOR_mod_wrap(
  HOR_mod_ls,
  sel_rows_tmp1 = CVhelp_dat_w,
  flength_in = 240,
  DOY_in = 1:250,
  years_in = NULL,
  route = "ORE"
)
}
\arguments{
\item{HOR_mod_ls}{named list with aic table and list of models}

\item{sel_rows_tmp1}{rows of 'CVhelp_dat_w' to use}

\item{flength_in}{fork length input scalar}

\item{DOY_in}{day of year subset}

\item{years_in}{years to subset}
}
\value{
dataframe with logit-scale predictions for survival model
}
\description{
Title
}
\details{
wrapper for generating HOR-TCJ survival predictions
}
