% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOR_TCJ_redef_fun.R
\name{HOR_TCJ_redef_fun}
\alias{HOR_TCJ_redef_fun}
\title{Title}
\usage{
HOR_TCJ_redef_fun(sel_rows_tmp1_in = sel_rows_tmp1, flength_in, YrRel_in = NA)
}
\arguments{
\item{sel_rows_tmp1_in}{rows of 'CVhelp_dat_w' to use}

\item{flength_in}{fork length input scalar}

\item{YrRel_in}{NA as default. Year and release group factor}
}
\value{
CVhelp_dat_w file with defined column names that match glmmTMB model.matrix so that predict.glmmTMB() can be used
}
\description{
Title
}
