% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_WYT_cols_vec.R
\name{mtch_WYT}
\alias{mtch_WYT}
\title{Title}
\usage{
mtch_WYT(intab = ann_HORbar_WYT_data)
}
\arguments{
\item{intab}{dataframe with WYT column for matching}
}
\value{
matching color associated with name attribut for color vector
}
\description{
Title
}
\examples{
WYT_cols <- utils_get_WYT_cols_vec()
WYT_cols[match(ann_HORbar_WYT_data$WYT,names(WYT_cols))]

}
