% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCJ_CHP_mod_wrap.R
\name{TCJ_CHP_mod_wrap}
\alias{TCJ_CHP_mod_wrap}
\title{Title}
\usage{
TCJ_CHP_mod_wrap(
  TCJ_CHP_pred_comp_ls_in = TCJ_CHP_pred_comp_ls,
  sel_rows_tmp1 = CVhelp_dat_w,
  flength_in = 240,
  DOY_in = 1:250,
  years_in = NULL,
  SJL_route_in = TRUE
)
}
\arguments{
\item{TCJ_CHP_pred_comp_ls_in}{}

\item{sel_rows_tmp1}{}

\item{flength_in}{}

\item{DOY_in}{}

\item{years_in}{}

\item{SJL_route_in}{}
}
\description{
Title
}
