% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_WYT_cols_vec.R
\name{utils_get_WYT_cols_vec}
\alias{utils_get_WYT_cols_vec}
\title{Title}
\usage{
utils_get_WYT_cols_vec()
}
\value{
of Water Year Type colors with name attributes
}
\description{
Title
}
\examples{
utils_get_WYT_cols_vec()

}
