% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ann_HORbar_WYT_data}
\alias{ann_HORbar_WYT_data}
\title{Central Valley Daily and Annual Environmental Data}
\format{
## `ann_HORbar_WYT_data`
A data frame with 14 rows and 6 columns:
\describe{
  \item{Year}{Calendar year}
  \item{barrier_days}{days with barrier}
  \item{barrier}{annual barrier status (i.e., any days with obstruction at all)}
  \item{ind}{index of 'wetness'}
  \item{wyt}{Water Year Type, abbreviation name}
  \item{WYT}{Water Year Type, full name}
  ...
}
}
\source{
<https://github.com/swhitCBR/CVhelp>
}
\usage{
ann_HORbar_WYT_data
}
\description{
A subset of data from multiple agencies with filtering and imputation carried about by the 'CVhelp" package
}
\keyword{datasets}
