% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_est_rch_surv_div_ui.R
\name{draw_est_rch_surv_div_ui}
\alias{draw_est_rch_surv_div_ui}
\title{Title}
\usage{
draw_est_rch_surv_div_ui(
  plots_in = list("HOR_TCJ_pred_ggpplt_rs1", "HOR_TCJ_pred_ggpplt_rs_2")
)
}
\arguments{
\item{plots_in}{}
}
\description{
Title
}
\examples{
draw_est_rch_surv_div_ui

}
