% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_est_rte_spec_surv_div_ui.R
\name{draw_est_rte_spec_surv_div_ui}
\alias{draw_est_rte_spec_surv_div_ui}
\title{Title}
\usage{
draw_est_rte_spec_surv_div_ui(
  plots_in = list("HOR_TCJ_pred_ggpplt_dup1b", "HOR_TCJ_pred_ggpplt_dup1a",
    "HOR_TCJ_pred_ggpplt_dup1e", "TCJ_CHPviaTRN_ggpplt")
)
}
\arguments{
\item{plots_in}{}
}
\description{
Title
}
\examples{
draw_est_rte_spec_surv_div_ui()

}
