% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_info_bttn_dropdown_ui.R
\name{draw_ibutt_dropdown_ui}
\alias{draw_ibutt_dropdown_ui}
\title{Title}
\usage{
draw_ibutt_dropdown_ui(
  inputId_in = "daily_var_def_ibutt2",
  ibox_content_label = "daily_values_box"
)
}
\arguments{
\item{inputId_in}{}
}
\description{
Title
}
