% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_ibutt_dropdown_ui.R
\name{draw_ibutt_dropdown_ui}
\alias{draw_ibutt_dropdown_ui}
\title{Title}
\usage{
draw_ibutt_dropdown_ui(
  inputId_in,
  ibox_content_label = "daily_values_box",
  width_in = "100\%"
)
}
\arguments{
\item{inputId_in}{}
}
\value{
\code{shiny.tag} object inside a \link[shinyWidgets:dropMenu]{shinyWidgets::dropMenu}
}
\description{
Title
}
