% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_inputs_box_ui.R
\name{draw_inputs_box_ui}
\alias{draw_inputs_box_ui}
\title{Title}
\usage{
draw_inputs_box_ui(
  inputs_panel_collapse = inputs_panel_collapse,
  init_data_source_in,
  collapsed_in = FALSE
)
}
\arguments{
\item{init_data_source_in}{data source object}

\item{input_tab_in}{input$tabs}
}
\value{
a \code{shiny.tag} object inside a `shinydashboardPlus::box()``
}
\description{
Title
}
\examples{
draw_inputs_box_ui(init_data_source="Previous years")

}
