% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_main_page_content_dynui.R
\name{draw_main_page_content_dynui}
\alias{draw_main_page_content_dynui}
\title{Title}
\usage{
draw_main_page_content_dynui(input_tab_in = input$tabs)
}
\arguments{
\item{input_tab_in}{}
}
\description{
Title
}
