% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sidebar_ui.R
\name{draw_sidebar_ui}
\alias{draw_sidebar_ui}
\title{Title}
\usage{
draw_sidebar_ui()
}
\value{
shinydashboard sidebar panel shinyUI code (i.e., within \code{shiny::tagList()})
}
\description{
Title
}
\details{
the
}
\examples{
draw_sidebar_ui()


}
\seealso{
calls \code{\link[=get_custom_css_txt]{get_custom_css_txt()}} for special formatting

Other {ui builders }: 
\code{\link{draw_start_loc_ui}()}

Other {front-end functions }: 
\code{\link{draw_start_loc_ui}()}
}
\concept{{front-end functions }}
\concept{{ui builders }}
