% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_start_loc_ui.R
\name{draw_start_loc_ui}
\alias{draw_start_loc_ui}
\title{Title}
\usage{
draw_start_loc_ui()
}
\arguments{
\item{loc_selected_in}{}
}
\value{
R shiny UI elements
}
\description{
Title
}
\examples{
draw_start_loc_ui()

}
\concept{{front-end functions }}
\concept{{ui builders }}
