% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_glmmTMB.R
\name{extract_glmmTMB_frame}
\alias{extract_glmmTMB_frame}
\title{Title}
\usage{
extract_glmmTMB_frame(glmmTMB_res_ls_in, mods_obj_nm = "HOR_TCJ_d2_mods")
}
\arguments{
\item{glmmTMB_res_ls_in}{named list containing glmmTMB model list and AIC table}

\item{mods_obj_nm}{name of the level-one element containing models}
}
\value{
list of dataframes used to fit glmmTMB models
}
\description{
Title
}
