% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOR_CHP_mod_wrap.R
\name{HOR_CHP_mod_wrap}
\alias{HOR_CHP_mod_wrap}
\title{Title}
\usage{
HOR_CHP_mod_wrap(
  HOR_CHP_mod_ls,
  sel_rows_tmp1 = CVhelp_dat_w,
  flength_in = 240,
  SJL_route_in = TRUE
)
}
\arguments{
\item{HOR_CHP_mod_ls}{named list with aic table and list of models}

\item{sel_rows_tmp1}{rows of 'CVhelp_dat_w' to use}

\item{flength_in}{fork length input scalar}
}
\value{
dataframe with logit-scale predictions for survival model
}
\description{
Title
}
\details{
wrapper for generating HOR-TCJ survival predictions
}
